Allow passing data directory via the command-line
This is a simple patch I use locally.  Requiring the data directory
directly above the executable is difficult especially when built as a
system pacakge.

Ideally this would use BSD's err.h for the err(3) API but I mostly kept
it in line with what is currently there and provided the fall back to
`../data` which should keep things working as usual.