Change the information window when a city tile is right-clicked.
Before it appeared horizontally centered, and either at the top or at the bottom of the screen.

With this change, the window appears with the top left corner at the mouse position, except if it would not fit in the window.

The new window positioning only works when the screen resolution is 1024x768 or higher.

The "special orders" menus have also been changed to appear depending on the information window position.