When creating a new building, set all its data to zero.
This prevents a bug which caused new buildings to have random values when a different city was loaded in the same playthrough.

For example, when I created new houses on empty maps and people moved in, the tents could appear with 90 pottery, 153 furniture and access to education and other services and would evolve to insulaes instantly.

On another occasion, a farm I had just build instantly created a cart pusher with 100 wheat.

This pull request seeks to fix those kinds of bugs.