Adds high DPI support for Julius.
Currently it's only enabled on Windows and Linux, as Mac OS doesn't seem to need scaling.

Reason for making the game high-DPI aware:

In windows and linux, when DPI was set to a custom value, even though the window scaled properly, the mouse cursor did not.
This resulted in a tiny, barely visible mouse when DPI was set very high.
Also, the right and bottom sides of the window did not scroll the map properly. Enabling high DPI support fixes that.

The application has full (per-monitor) DPI-suport in Windows, and system DPI support in Linux, automatically resizing when the DPI is changed.

Mac OS X is a whole can of worms regarding DPI, because it requires the application to be bundled as an app in order to support dpi-awareness.

Even though that is not hard to do (it only requires some changes to the CMakeLists.txt file), bundling an app makes it impossible to just dump the executable on the same folder of the data files and start playing the game.
Also, since SDL scales everything properly (including the mouse cursor) in Mac OS, there is no need for now to implement DPI awareness.