Change the way window invalidation works. (#42)
There are now two types of invalidation:
- An immediate invalidation, which prevents any further ticks from occuring until the screen is refreshed;
- An invalidation "on draw", which allows the remaining game ticks to be processed before redrawing the screen.

Reasoning for the change:

With the old code, there was a bug with window drawing when game speed was set to >100%.
Specifically, the victory screen did not appear and instead the victory video showed immediately (with no chance to continue governing).

The bug was due to the fact that when victory condition was met, a value that the victory popup was shown was set to true.
After that, the screen was invalidated to draw the victory dialog.
However, since the game processed further ticks, it went back to the victory condiction routine (city_victory_check in victory.c) and, as the popup was set to already having been shown, the game loaded the victory video without ever loading the victory popup.

By setting two different timings to a redraw request (either after the current tick or after all the remaning ticks have been processed in game_run), the game is able to immediately show popup windows and prevent the city to keep running in the background.
Also, certain elements, like tooltips and the sidebar scrolling, do not need to interrupt the ticks, hence they are only redrawn after all ticks have been processed. If that wasn't done, game speed would be reduced to 1 tick per frame, slowing down the game speed whenever a tooltip was being displayed of the sidebar scrolled.