Fix soldiers not acting properly on enormous maps
The problem with formation_position_x/y is that the variable is used as unsigned int for the player's soldiers, and as a signed int for enemy soldiers. This leads to integer wrapping. Using unsigned char does work for the enemy soldiers, but that's pure coincidence. Introducing a union makes the difference explicit and seemed the only future-proof way to me.
Issue #69, supersedes PR #78