Cache the minmap to improve performance on window invalidation.
As a consequence of this change, two bugs were found and fixed.

When the minimap was invalidated (data.minimap_redraw_requested set to 1 in sidebar.c), it was never set back to zero so the minimap was being redrawn every single frame.

graphics_save_from_buffer and graphics_copy_from_buffer were using the height size, and not the width size, to find the starting pixel.