Reduce codecov build failures.
Currently a build fails if either the tested coverage of the PR's whole build and of the patch is below the master's coverage percentage.

That seems to strict, as some patches may be added to areas that weren't covered so they will alsays stick to 0% coverage, reduce the whole bild coverage and therefore fail the build.

This changes that so the patches aren't checked and codecov only fails if the whole build's coverage falls more than 0.5%.