Improve the visual feedback when clearing land.
If the "clear land" tool is selected, the tile the mouse is hovered on
will be highlighted in red. If that tile is a building, the whole
building will be highlighted.

If the "clear land" tool is dragged, the area to be cleared will be
highlighted in red. If the area includes buildings, the entire building
will be highlighted.