Implement text display limits to text input fields.
If the text written to an input is larger than its input field width,
only the text that fits to the width of the input will be shown.

Which portion of the text will be shown depends on the cursor position.

If there is text input to the left or right that is not shown, a small
arrow will appear to indicate there is further text.

So far this is only used in the file load/save/delete dialog.

This fixes the visual glitches when using long file names for cities.