Make game speed independent of fps
Before, at speeds at or above 100%, the game played at 1 (or 2 at 200$, 3 at 300%, etc...) in-game ticks per frame. This caused 100% game speed to be much faster on 144hz monitors and slower when fps was low.

This PR fixes that, and now the game always runs at the same speed regardless of fps, with a maximum limit of 20 ticks per frame.