Refactor platform code
The aim of this commit is to further merge Vita/Switch code with base Julius code.

Add platform-dependant callbacks on game init and for each frame instead of using a specific ifdef for each platform.

Refactor virtual keyboard code, allowing Vita and Switch keyboards to automatically appear when pressing the textbox. It also allows the current text to appear when the textbox is clicked.