Prevent to build multiple unique buildings by using undo
If you removed an unique building (e.g. an hippodrome) then hit the undo button, you were able to build a new one, as it was not added back to the city data during the undo.

I have omitted the monuments as I did not manage to undo a monument deletion.