Calculate favour changes on actual salary paid
In Caesar 3, favour changes depending on the salary are only based on what your salary is set in January, allowing to easily bypass the intended penalty.

This commit changes it, so the rank used for favour calculations is the lowest rank that could receive the salary the player has paid himself.