Fix build images in the sidebar
Since the last sidebar refactor, every building would show the "housing" build image in the sidebar due to the selected_submenu variable being reset to NONE. Reinstate the old behaviour of showing the correct image when a building is selected for construction. This implementation can also be used for buildings selected using hotkeys.
Issue #620